+++ src/components/Heading/Heading.stories.tsx
@@ -1,9 +1,7 @@
-import { BADGE } from '@geometricpanda/storybook-addon-badges';
 import type { StoryObj, Meta } from '@storybook/react';
 import React from 'react';
 
-import { Heading, VARIANTS } from './Heading';
-import styles from './Heading.stories.module.css';
+import { Heading } from './Heading';
 
 export default {
   title: 'Components/Heading',
@@ -14,111 +12,103 @@ export default {
 } as Meta<Args>;
 
 type Args = React.ComponentProps<typeof Heading>;
+type Story = StoryObj<Args>;
 
-export const Heading1: StoryObj<Args> = {
+export const Heading1: Story = {
   args: {
     size: 'h1',
+    preset: 'headline-lg',
     children: 'Heading 1',
   },
 };
 
-export const Heading2: StoryObj<Args> = {
+export const Heading2: Story = {
   args: {
     size: 'h2',
+    preset: 'headline-md',
     children: 'Heading 2',
   },
 };
 
-export const Heading3: StoryObj<Args> = {
+export const Heading3: Story = {
   args: {
     size: 'h3',
+    preset: 'headline-sm',
     children: 'Heading 3',
   },
 };
 
-export const Heading4: StoryObj<Args> = {
+export const Heading4: Story = {
   args: {
     size: 'h4',
+    preset: 'title-md',
     children: 'Heading 4',
   },
 };
 
-export const Heading5: StoryObj<Args> = {
+export const Heading5: Story = {
   args: {
     size: 'h5',
+    preset: 'title-sm',
     children: 'Heading 5',
   },
 };
 
-export const Heading6: StoryObj<Args> = {
+export const Heading6: Story = {
   args: {
     size: 'h6',
+    preset: 'title-xs',
     children: 'Heading 6',
-  },
-  parameters: {
-    badges: [BADGE.DEPRECATED],
   },
 };
 
-export const Heading7: StoryObj<Args> = {
+export const Heading7: Story = {
   args: {
     size: 'h7',
     children: 'Heading 7',
   },
-  parameters: {
-    badges: [BADGE.DEPRECATED],
-  },
 };
 
-export const TitleXs: StoryObj<Args> = {
-  args: {
-    size: 'title-xs',
-    children: 'Title xs heading size',
-  },
-};
-
-export const Variants: StoryObj<Args> = {
-  ...Heading1,
-  render: (args) => {
-    const headings: React.ReactNode[] = [];
-    VARIANTS.forEach((variant) => {
-      if (variant !== 'info')
-        headings.push(
-          <Heading {...args} key={`${variant}`} variant={variant}>
-            {variant}
-          </Heading>,
-        );
-    });
-    return (
-      <div className={styles['variant__table']}>
-        <div className={styles['variant--white']}>{headings}</div>
-        <div className={styles['variant--light']}>{headings}</div>
-        <div className={styles['variant--dark']}>{headings}</div>
-      </div>
-    );
-  },
-  parameters: {
-    /**
-     * Mainly for visual use and other stories generate enough confidence for our needs.
-     */
-    axe: {
-      skip: true,
-    },
-    snapshot: { skip: true },
-  },
-};
-
-export const Sizes: StoryObj<Args> = {
-  render: () => {
-    const sizes = ['h1', 'h2', 'h3', 'h4', 'h5', 'h6'] as const;
-    const combinations = sizes.map((as: (typeof sizes)[number]) => {
-      const headings = sizes.map((size: (typeof sizes)[number]) => (
-        <Heading as={as} key={`as${as}size${size}`} size={size}>
-          as: {as} size: {size}
-        </Heading>
-      ));
-      return <div key={as}>{headings}</div>;
-    });
-    return <div className={styles['variant__table']}>{combinations}</div>;
-  },
+/**
+ * Here we demonstrate how to use utility classes to augment the headings.
+ * Note that when present, `preset` will override the deprecated props.
+ */
+export const UsingColorTokens: Story = {
+  render: (args) => (
+    <div>
+      <Heading
+        {...args}
+        className="text-utility-warning"
+        preset="title-md"
+        size="h1"
+      >
+        using <code>text-utility-warning</code> utility class
+      </Heading>
+      <Heading
+        {...args}
+        as="h2"
+        className="text-utility-success"
+        preset="title-md"
+        size="h3"
+      >
+        using <code>text-utility-success</code> utility class
+      </Heading>
+      <Heading
+        {...args}
+        className="text-utility-error"
+        preset="title-md"
+        size="h3"
+        variant="white"
+      >
+        using <code>text-utility-error</code> utility class
+      </Heading>
+      <Heading
+        preset="title-md"
+        size="h4"
+        style={{ color: 'var(--eds-theme-color-text-utility-success)' }}
+      >
+        using inline color
+      </Heading>
+    </div>
+  ),
 };